+++ packages/ispect/lib/src/features/inspector/inspector.dart
@@ -1,1 +1,3 @@
+export 'package:draggable_panel/draggable_panel.dart';
+
 export './src/inspector.dart';