+++ packages/ispect/lib/src/features/inspector/src/inspector.dart
@@ -3,6 +3,7 @@
 import 'dart:async';
 import 'dart:ui' as ui;
 
+import 'package:draggable_panel/draggable_panel.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
@@ -18,7 +19,6 @@ import 'package:ispect/src/features/inspector/src/widgets/color_picker/utils.dar
 import 'package:ispect/src/features/inspector/src/widgets/inspector/box_info.dart';
 import 'package:ispect/src/features/inspector/src/widgets/inspector/overlay.dart';
 import 'package:ispect/src/features/inspector/src/widgets/multi_value_listenable.dart';
-import 'package:ispect/src/features/inspector/src/widgets/panel/ispect_panel.dart';
 import 'package:ispect/src/features/inspector/src/widgets/zoomable_color_picker/overlay.dart';
 import 'package:ispect/src/features/snapshot/feedback_plus.dart';
 import 'package:share_plus/share_plus.dart';
@@ -117,8 +117,7 @@ class Inspector extends StatefulWidget {
     ]);
   }
 
-  static InspectorState? maybeOf(BuildContext? context) =>
-      context?.findAncestorStateOfType<InspectorState>();
+  static InspectorState? maybeOf(BuildContext? context) => context?.findAncestorStateOfType<InspectorState>();
 
   @override
   InspectorState createState() => InspectorState();
@@ -128,8 +127,7 @@ class InspectorState extends State<Inspector> {
   bool _isPanelVisible = false;
   bool get isPanelVisible => _isPanelVisible;
 
-  void togglePanelVisibility() =>
-      setState(() => _isPanelVisible = !_isPanelVisible);
+  void togglePanelVisibility() => setState(() => _isPanelVisible = !_isPanelVisible);
 
   final _stackKey = GlobalKey();
   final _repaintBoundaryKey = GlobalKey();
@@ -203,9 +201,7 @@ class InspectorState extends State<Inspector> {
 
     if (boxes.isEmpty) return;
 
-    final overlayOffset =
-        (_stackKey.currentContext!.findRenderObject()! as RenderStack)
-            .localToGlobal(Offset.zero);
+    final overlayOffset = (_stackKey.currentContext!.findRenderObject()! as RenderStack).localToGlobal(Offset.zero);
 
     _currentRenderBoxNotifier.value = BoxInfo.fromHitTestResults(
       boxes,
@@ -293,8 +289,7 @@ class InspectorState extends State<Inspector> {
 
   Future<void> _extractByteData() async {
     if (_image != null) return;
-    final boundary = _repaintBoundaryKey.currentContext!.findRenderObject()!
-        as RenderRepaintBoundary;
+    final boundary = _repaintBoundaryKey.currentContext!.findRenderObject()! as RenderRepaintBoundary;
 
     final pixelRatio = MediaQuery.devicePixelRatioOf(context);
 
@@ -305,9 +300,8 @@ class InspectorState extends State<Inspector> {
   Offset _extractShiftedOffset(Offset offset) {
     final pixelRatio = MediaQuery.devicePixelRatioOf(context);
 
-    var offset0 = (_repaintBoundaryKey.currentContext!.findRenderObject()!
-            as RenderRepaintBoundary)
-        .globalToLocal(offset);
+    var offset0 =
+        (_repaintBoundaryKey.currentContext!.findRenderObject()! as RenderRepaintBoundary).globalToLocal(offset);
 
     // ignore: join_return_with_assignment
     offset0 *= pixelRatio;
@@ -320,9 +314,7 @@ class InspectorState extends State<Inspector> {
 
     final shiftedOffset = _extractShiftedOffset(offset);
 
-    final overlayOffset =
-        (_stackKey.currentContext!.findRenderObject()! as RenderStack)
-            .localToGlobal(Offset.zero);
+    final overlayOffset = (_stackKey.currentContext!.findRenderObject()! as RenderStack).localToGlobal(Offset.zero);
 
     _zoomImageOffsetNotifier.value = shiftedOffset;
     _zoomOverlayOffsetNotifier.value = offset - overlayOffset;
@@ -330,8 +322,7 @@ class InspectorState extends State<Inspector> {
 
   void _onPointerScroll(PointerScrollEvent scrollEvent) {
     if (_zoomStateNotifier.value) {
-      final newValue =
-          _zoomScaleNotifier.value + 1.0 * -scrollEvent.scrollDelta.dy.sign;
+      final newValue = _zoomScaleNotifier.value + 1.0 * -scrollEvent.scrollDelta.dy.sign;
 
       if (newValue < 1.0) {
         return;
@@ -397,8 +388,7 @@ class InspectorState extends State<Inspector> {
             builder: (_) {
               final child = widget.child;
 
-              final isAbsorbingPointer =
-                  _inspectorStateNotifier.value || _zoomStateNotifier.value;
+              final isAbsorbingPointer = _inspectorStateNotifier.value || _zoomStateNotifier.value;
 
               return Listener(
                 behavior: HitTestBehavior.translucent,
@@ -499,48 +489,43 @@ class InspectorState extends State<Inspector> {
                 _zoomStateNotifier,
                 _byteDataStateNotifier,
               ],
-              builder: (context) => ISpectMenuPanel(
+              builder: (context) => DraggablePanel(
                 borderRadius: const BorderRadius.all(Radius.circular(16)),
-                panelShape: PanelShape.rectangle,
-                observer: widget.observer,
                 backgroundColor: context.isDarkMode
                     ? context.ispectTheme.colorScheme.primaryContainer
                     : context.ispectTheme.colorScheme.primary,
                 initialPosition: widget.initialPosition,
-                onPositionChanged: (x, y) =>
-                    widget.onPositionChanged?.call(x, y),
+                onPositionChanged: (x, y) => widget.onPositionChanged?.call(x, y),
                 items: [
-                  ISpectPanelItem(
-                    icon: _controller.inLoggerPage
-                        ? Icons.undo_rounded
-                        : Icons.reorder_rounded,
+                  (
+                    icon: _controller.inLoggerPage ? Icons.undo_rounded : Icons.reorder_rounded,
                     enableBadge: _controller.inLoggerPage,
                     onTap: (_) {
                       _launchInfospect(context);
                     },
                   ),
-                  ISpectPanelItem(
+                  (
                     icon: Icons.monitor_heart_outlined,
                     enableBadge: iSpect.isPerformanceTrackingEnabled,
                     onTap: (_) {
                       iSpect.togglePerformanceTracking();
                     },
                   ),
-                  ISpectPanelItem(
+                  (
                     icon: Icons.format_shapes_rounded,
                     enableBadge: _inspectorStateNotifier.value,
                     onTap: (_) {
                       _onInspectorStateChanged(!_inspectorStateNotifier.value);
                     },
                   ),
-                  ISpectPanelItem(
+                  (
                     icon: Icons.colorize_rounded,
                     enableBadge: _zoomStateNotifier.value,
                     onTap: (_) {
                       _onZoomStateChanged(!_zoomStateNotifier.value);
                     },
                   ),
-                  ISpectPanelItem(
+                  (
                     icon: Icons.camera_alt_rounded,
                     enableBadge: feedback.isVisible,
                     onTap: (_) {
@@ -550,6 +535,7 @@ class InspectorState extends State<Inspector> {
                   ...widget.options.panelItems,
                 ],
                 buttons: widget.options.panelButtons,
+                child: null,
               ),
             ),
           ),
@@ -560,8 +546,7 @@ class InspectorState extends State<Inspector> {
   void _toggleFeedback(FeedbackController feedback, BuildContext context) {
     if (!feedback.isVisible) {
       feedback.show((feedback) async {
-        final screenshotFilePath =
-            await writeImageToStorage(feedback.screenshot);
+        final screenshotFilePath = await writeImageToStorage(feedback.screenshot);
 
         await Share.shareXFiles(
           [screenshotFilePath],